Revamp Authentication System

+++ PokemonGo.py
@@ -1,18 +1,22 @@
-import os.path
 from tools import utils, api_tools as api, poke_utils as poke
 from tools.log_tool import get_logger
-import logging as log
 
-from google.auth.transport.requests import Request
-from google.oauth2.credentials import Credentials
-from google_auth_oauthlib.flow import InstalledAppFlow
 from googleapiclient.discovery import build
 
 SCOPES = ["https://www.googleapis.com/auth/calendar"]
 
-def main(): 
+def pokemon_go(creds, delete_future_events:bool): 
     """Script for Pokemon Go Events"""
-    
+    # Define some variables
+    api_url = "https://raw.githubusercontent.com/bigfoott/ScrapedDuck/data/events.json"
+    cache_file = "json/pokemonGo_events.json"
+    whiteList = ["community-day", "event", 
+                 "live-event", "pokemon-go-fest", "pokemon-spotlight-hour", 
+                 "season", "pokemon-go-tour", "raid-day", "elite-raids", 
+                 "raid-battles", "raid-hour", "raid-weekend", "go-battle-league"] # Make sure to edit whiteList in all_day_event() if have to
+
+    log = get_logger()
+    log.info("--------------START of PokemonGo Script-------------")
     # Fetch the data from the API
     data = api.fetch_api_data(api_url)
     
@@ -27,7 +31,8 @@ def main():
     calendarID = "1c565d24c2befe1ade79037bc085ab3b72fc1860cc463f6547d2e585253e26cf@group.calendar.google.com"
     
     # Do you want to delete all future events in the calendar? 
-    utils.delete_future_events(service, calendarID)
+    if delete_future_events:
+        utils.delete_future_events(service, calendarID)
       
     # Go through each event from json file
     for event in api.load_data_from_file(cache_file):      
@@ -38,40 +43,4 @@ def main():
         # Adds event to calendar
         utils.add_events(service, calendarID, new_event, duplicate_check=False) 
         
-    log.info("-----------------END of PokemonGo Script-----------------\n")
-            
-                
-if __name__ == "__main__":
-    # Call logging to start
-    log = get_logger()
-    log.info("-----------------START of PokemonGo Script-----------------")
-    
-    """Handles all of Google OAuth"""
-    creds = None
-    # The file token.json stores the user's access and refresh tokens, and is
-    # created automatically when the authorization flow completes for the first
-    # time.
-    if os.path.exists("json/token.json"):
-        creds = Credentials.from_authorized_user_file("json/token.json", SCOPES)
-
-    # If there are no (valid) credentials available, let the user log in.
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file("json/credentials.json", SCOPES)
-            creds = flow.run_local_server(port=0)
-    
-        # Save the credentials for the next run
-        with open("json/token.json", "w") as token:
-            token.write(creds.to_json())
-            
-    # Define some variables
-    api_url = "https://raw.githubusercontent.com/bigfoott/ScrapedDuck/data/events.json"
-    cache_file = "json/pokemonGo_events.json"
-    whiteList = ["community-day", "event", 
-                 "live-event", "pokemon-go-fest", "pokemon-spotlight-hour", 
-                 "season", "pokemon-go-tour", "raid-day", "elite-raids", 
-                 "raid-battles", "raid-hour", "raid-weekend", "go-battle-league"] # Make sure to edit whiteList in all_day_event() if have to
-    
-    main()
+    log.info("-------------END of PokemonGo Script-------------\n")
